fix: add a grpc retry policy for the cloud engine (#45)
What:
On the initial session set up, resources might take a while to come up,
so we need to wait for them to be available.
Grpc provides a mechanism to do automated retries (by passing a policy).
This adds a default policy for 5 exponential retries with the initial
one waiting 10 seconds.

Tests:
Unit Tests
Tested connecting to the cloud and forcing a new session to be created
(by closing the previous session). No errors during connection.