fix: use basic types for expressions (#48)
**What**
For semantic classify we can pass either an enum or a list of strings
that will be transformed into an enum. When the list of strings is
passed, that enum is dynamically generated, causing problems when
calling pickle to serialize the plan.
For semantic extract we can pass either a ExtractSchema or a pydantic
model, in the case of a ExtractSchema, this is transformed into a
pydantic model during runtime causing a problem while pickling the plan.

**The Fix**
The proposal is to only do the transformations during execution (we
might do transformations to validate), but while the logical plan is
generated there shouldn't be any dynamic calls.

---------

Co-authored-by: Brandon Callender <bcallender@users.noreply.github.com>