feat: Support for persistent views (#41)
This PR adds the support to save a view of a dataframe. With this we are
able to persist a view of the dataframe without the actual execution of
the sources. This includes APIs on saving a dataframe as a view, and
also on creating/describing/dropping views. This also modifies the
`schema_storage` file by renaming it to `system_table_client`, and now
also includes a view metadata table that stores the view data.

---------

Co-authored-by: German Gil <german@typedef.ai>
Co-authored-by: German Gil <169086731+german-typedef@users.noreply.github.com>