docs: Update example notebooks (#144)
I'm updating the notebooks in our examples to add some CTAs, links to
other examples from within the notebooks and most importantly to make
the notebooks possible to directly run inside Google colab.

I'm also updating the README.md to add a link to open the notebook in
Google colab.