fix: raise exception in user thread instead of shutting down when request above model tpm (#228)
## What changed?

https://github.com/typedef-ai/fenic/pull/210/files fixes a bug where we
hang forever if a single request token estimate is above the model's tpm
limit, but it does so by incorrectly shutting down the processing queue.
This fix patches it so we raise the exception in the user thread
instead.

## How to test?

Run unit tests with semantic operations, set model configuration with
low tpm (like 10), make sure the semantic operations raise immediately